Fix heroku ws app (#302)
* Improve 3.2

* Fix heroku app

* Get rid of main

* Update notebook

* Fix websocket_app.py

* args.gateway_url -> gateway_url

* Change Model ids (db.Integer -> db.BigInteger)

* Update notebook

* Remove outdated comment

+++ app/websocket/app/main/persistence/models.py
@@ -49,7 +49,7 @@ class WorkerObject(db.Model):
 
     __tablename__ = "worker_objects"
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     description = db.Column(db.String(64), index=True)
     data = db.Column(db.LargeBinary(128))
     worker_id = db.Column(db.String(64), db.ForeignKey("workers.id"))
@@ -74,7 +74,7 @@ class TorchTensor(db.Model):
 
     __tablename__ = "torch_tensors"
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     data = db.Column(db.LargeBinary(128))
 
     @property
